refactor: remove barrel files
