+++ src/config/authInfoConfig.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { AuthFields } from '../org';
+import { AuthFields } from '../org/authInfo';
 import { ConfigFile } from './configFile';
 
 /**
