--- src/org/index.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2020, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-
-export * from './authInfo';
-export * from './authRemover';
-export * from './connection';
-export * from './org';
-export * from './permissionSetAssignment';
-export * from './user';
