--- src/stateAggregator/index.ts
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2021, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-
-export * from './accessors/orgAccessor';
-export * from './accessors/aliasAccessor';
-export * from './stateAggregator';
