+++ src/testSetup.ts
@@ -39,8 +39,11 @@ import { SfError } from './sfError';
 import { SfProject, SfProjectJson } from './sfProject';
 import * as aliasAccessorEntireFile from './stateAggregator/accessors/aliasAccessor';
 import { CometClient, CometSubscription, Message, StreamingExtension } from './status/streamingClient';
-import { OrgAccessor, StateAggregator } from './stateAggregator';
-import { AuthFields, Org, SandboxFields, User, UserFields } from './org';
+import { StateAggregator } from './stateAggregator/stateAggregator';
+import { OrgAccessor } from './stateAggregator/accessors/orgAccessor';
+import { Org, SandboxFields } from './org/org';
+import { AuthFields } from './org/authInfo';
+import { User, UserFields } from './org/user';
 import { SandboxAccessor } from './stateAggregator/accessors/sandboxAccessor';
 import { Global } from './global';
 import { uniqid } from './util/uniqid';
