+++ test/unit/org/scratchOrgCreateTest.ts
@@ -8,7 +8,8 @@ import { expect } from 'chai';
 import * as sinon from 'sinon';
 import { Duration } from '@salesforce/kit';
 import { Connection } from 'jsforce';
-import { AuthInfo, Org } from '../../../src/org';
+import { Org } from '../../../src/org/org';
+import { AuthInfo } from '../../../src/org/authInfo';
 import { SfProject, SfProjectJson } from '../../../src/sfProject';
 import { scratchOrgCreate, ScratchOrgCreateOptions, scratchOrgResume } from '../../../src/org/scratchOrgCreate';
 import { ScratchOrgCache } from '../../../src/org/scratchOrgCache';
