+++ test/unit/org/scratchOrgInfoApiTest.ts
@@ -10,7 +10,9 @@ import * as sinon from 'sinon';
 import { assert, expect } from 'chai';
 import { Duration, env } from '@salesforce/kit';
 import { stubMethod } from '@salesforce/ts-sinon';
-import { AuthInfo, Connection, Org } from '../../../src/org';
+import { Org } from '../../../src/org/org';
+import { Connection } from '../../../src/org/connection';
+import { AuthInfo } from '../../../src/org/authInfo';
 import { shouldThrow } from '../../../src/testSetup';
 import { MyDomainResolver } from '../../../src/status/myDomainResolver';
 import SettingsGenerator from '../../../src/org/scratchOrgSettingsGenerator';
