+++ test/unit/org/userTest.ts
@@ -9,7 +9,11 @@ import { AnyJson, isString } from '@salesforce/ts-types';
 import { expect } from 'chai';
 import { SecureBuffer } from '../../../src/crypto/secureBuffer';
 import { MockTestOrgData, shouldThrow, shouldThrowSync, TestContext } from '../../../src/testSetup';
-import { DefaultUserFields, User, PermissionSetAssignment, Org, Connection, AuthInfo } from '../../../src/org';
+import { Org } from '../../../src/org/org';
+import { DefaultUserFields, User } from '../../../src/org/user';
+import { AuthInfo } from '../../../src/org/authInfo';
+import { Connection } from '../../../src/org/connection';
+import { PermissionSetAssignment } from '../../../src/org/permissionSetAssignment';
 
 describe('User Tests', () => {
   const $$ = new TestContext();
