+++ test/unit/stateAggregator/accessors/aliasAccessorTest.ts
@@ -9,7 +9,8 @@ import { existsSync } from 'node:fs';
 import { rm } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import { expect } from 'chai';
-import { FILENAME, StateAggregator } from '../../../../src/stateAggregator';
+import { StateAggregator } from '../../../../src/stateAggregator/stateAggregator';
+import { FILENAME } from '../../../../src/stateAggregator/accessors/aliasAccessor';
 import { MockTestOrgData, TestContext } from '../../../../src/testSetup';
 import { Global } from '../../../../src/global';
 import { uniqid } from '../../../../src/util/uniqid';
