+++ test/unit/stateAggregator/accessors/orgAccessorTest.ts
@@ -5,8 +5,8 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { expect } from 'chai';
-import { StateAggregator } from '../../../../src/stateAggregator';
-import { AuthFields } from '../../../../src/org';
+import { StateAggregator } from '../../../../src/stateAggregator/stateAggregator';
+import { AuthFields } from '../../../../src/org/authInfo';
 import { MockTestOrgData, shouldThrowSync, TestContext } from '../../../../src/testSetup';
 import { uniqid } from '../../../../src/util/uniqid';
