+++ test/unit/status/streamingClientTest.ts
@@ -10,7 +10,7 @@ import { SinonSpyCall } from 'sinon';
 import { Duration } from '@salesforce/kit';
 import { get, JsonMap } from '@salesforce/ts-types';
 import { CometClient, StatusResult, StreamingClient } from '../../../src/status/streamingClient';
-import { Connection } from '../../../src/org';
+import { Connection } from '../../../src/org/connection';
 // import { Crypto } from '../../../src/crypto/crypto';
 import { Org } from '../../../src/org/org';
 import { SfError } from '../../../src/sfError';
